+++ c/src/lib/libcpu/sh/sh7045/sci/sci_termios.c
@@ -355,10 +355,10 @@ int sh_sci_last_close(
 /*
  * Interrupt aware write routine
  */
-int sh_sci_write_support_int(
+ssize_t sh_sci_write_support_int(
   int            minor,
   const char    *buf,
-  int            len
+  size_t         len
 )
 {
     if (!len)
@@ -379,10 +379,10 @@ int sh_sci_write_support_int(
 /*
  * Polled write method
  */
-int sh_sci_write_support_polled(
+ssize_t sh_sci_write_support_polled(
   int            minor,
   const char    *buf,
-  int            len
+  size_t         len
 )
 {
     int count = 0;
